public method to create okHttpClient
